requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pdfplumber
python-docx
faiss-cpu
numpy
scikit-learn
